+++ lib/magic_auth/magic_auth_test.ex
@@ -1,6 +1,11 @@
 defmodule MagicAuth.TestHelpers do
   def log_in_session(conn, params) do
-    session = MagicAuth.create_session!(params)
+    session =
+      conn
+      |> MagicAuth.session_metadata()
+      |> Map.merge(params)
+      |> MagicAuth.create_session!()
+
     MagicAuth.put_token_in_session(conn, session.token)
   end
 end